Corrections diverses front
