+++ front/src/app/map/map.component.ts
@@ -5,7 +5,6 @@ import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
 import {timer} from 'rxjs';
 import {Position} from "../position";
 import {MapboxService} from "./mapbox.service";
-import {MapService} from "./map.service";
 
 @Component({
   selector: 'app-map',
@@ -13,6 +12,10 @@ import {MapService} from "./map.service";
   styleUrls: ['./map.component.scss']
 })
 export class MapComponent implements OnInit {
+
+  private readonly TIMER_GEOLOCALISATION = 2000;
+
+  private readonly TIMER_FLYING_MAP = 3000;
 
   private rechercheDestination?: boolean;
 
@@ -24,13 +27,16 @@ export class MapComponent implements OnInit {
 
   public geolocalisation!: mapboxgl.GeolocateControl;
 
-  private readonly TIMER_GEOLOCALISATION = 2000;
+  public positionDepart!: Position;
 
-  private readonly TIMER_FLYING_MAP = 3000;
+  public positionArrivee?: Position;
+  
+  public dureeTrajet?: string;
+
+  public instructionTrajet?: string;
 
   constructor(private stationService: StationService,
-              private mapboxService:MapboxService,
-              public mapService: MapService) {
+              public mapboxService: MapboxService) {
   }
 
   ngOnInit(): void {
@@ -49,8 +55,11 @@ export class MapComponent implements OnInit {
   }
 
   public affichageTrajetStationLaPlusProche(): void {
-    this.stationService.findStationLaPlusProche(this.mapService.coordsStart).subscribe(station => {
-      this.mapService.coordsEnd = [station.positionDto.longitude, station.positionDto.latitude];
+    this.stationService.findStationLaPlusProche(this.positionDepart).subscribe(station => {
+      this.positionArrivee = {
+        longitude: station.positionDto.longitude,
+        latitude: station.positionDto.latitude,
+      };
       this.creerRoute();
     })
   }
@@ -58,7 +67,7 @@ export class MapComponent implements OnInit {
   public setCoordsOfStartToPosition(): void {
     this.geolocalisation.trigger();
     timer(this.TIMER_GEOLOCALISATION).subscribe(() => {
-      this.mapService.coordsStart = {
+      this.positionDepart = {
         longitude: this.map.getCenter().lng,
         latitude: this.map.getCenter().lat
       }
@@ -70,15 +79,18 @@ export class MapComponent implements OnInit {
     this.rechercheDestination = true;
     this.geolocalisation.trigger();
     timer(this.TIMER_GEOLOCALISATION).subscribe(() => {
-      this.mapService.coordsEnd = [this.map.getCenter().lng, this.map.getCenter().lat]
+      this.positionArrivee = {
+        longitude: this.map.getCenter().lng,
+        latitude: this.map.getCenter().lat,
+      };
       this.isPositionOk = true;
     });
   }
 
   private creerRoute(): void {
-    this.mapService.removeRoute(this.map);
+    this.mapboxService.removeRoute(this.map);
     this.afficherDureeTrajet = true;
-    this.mapService.getRouteInformations(this.map)
+    this.calculerInformationsItineraire();
   }
 
 
@@ -102,7 +114,10 @@ export class MapComponent implements OnInit {
         if (rechercheTrajet?.checked) {
           this.rechercheDestination = true;
           new mapboxgl.Marker().setLngLat([this.map.getCenter().lng, this.map.getCenter().lat]).addTo(this.map);
-          this.mapService.coordsEnd = [this.map.getCenter().lng, this.map.getCenter().lat];
+          this.positionArrivee = {
+            longitude: this.map.getCenter().lng,
+            latitude: this.map.getCenter().lat,
+          };
           this.creerRoute();
         }
       })
@@ -114,7 +129,7 @@ export class MapComponent implements OnInit {
 
     document.querySelector('#rechercheDepart')?.appendChild(rechercheDepart.onAdd(this.map));
     timer(this.TIMER_FLYING_MAP).subscribe(() => {
-      this.mapService.coordsStart = {
+      this.positionDepart = {
         longitude: this.map.getCenter().lng,
         latitude: this.map.getCenter().lat
       }
@@ -127,7 +142,7 @@ export class MapComponent implements OnInit {
         positionRecherche = new mapboxgl.Marker().setLngLat([this.map.getCenter().lng, this.map.getCenter().lat])
         positionRecherche.addTo(this.map)
         if (trajet?.checked) {
-          this.mapService.coordsStart = {
+          this.positionDepart = {
             longitude: this.map.getCenter().lng,
             latitude: this.map.getCenter().lat
           }
@@ -162,5 +177,26 @@ export class MapComponent implements OnInit {
 
   private infosStations(capacity: number, placeDispo: number): string {
     return 'capacité : ' + capacity + '<br>' + 'places disponibles : ' + placeDispo
+  }
+
+  private calculerInformationsItineraire() {
+    let xmlHttpRequest = new XMLHttpRequest();
+    xmlHttpRequest.responseType = 'json';
+    xmlHttpRequest.open('GET', this.mapboxService.buildInformationsItineraireUrl(this.positionDepart, this.positionArrivee), true);
+    xmlHttpRequest.onload = () => {
+      let apiDirectionResponse = xmlHttpRequest.response;
+      let duration = apiDirectionResponse.routes[0].duration / 60;
+      let stepsOfRoad = apiDirectionResponse.routes[0].legs[0].steps;
+      let coords = apiDirectionResponse.routes[0].geometry;
+      let instructionsTrajet = '';
+      stepsOfRoad.forEach(function (step: any) {
+        instructionsTrajet += step.maneuver.instruction;
+      });
+      this.instructionTrajet = instructionsTrajet;
+
+      this.dureeTrajet = duration.toFixed(2)
+      this.mapboxService.dessinerItineraire(this.map, coords);
+    };
+    xmlHttpRequest.send();
   }
 }
