--- front/src/app/map/map.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { MapService } from './map.service';
-
-describe('MapService', () => {
-  let service: MapService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(MapService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
